Refactored routes
* my tickets
